update self update
