+++ global/update/update_windows.go
@@ -1,0 +1,35 @@
+package update
+
+import (
+	"archive/zip"
+	"bytes"
+	"fmt"
+	"io"
+	"net/http"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func Update(url string) {
+	resp, err := http.Get(url)
+	if err != nil {
+		log.Error("更新失败: ", err)
+		return
+	}
+	defer resp.Body.Close()
+	wc := WriteCounter{}
+	rsp, _ := io.ReadAll(io.TeeReader(resp.Body, &wc))
+	reader, _ := zip.NewReader(bytes.NewReader(rsp), resp.ContentLength)
+	file, err := reader.Open("go-cqhttp.exe")
+	if err != nil {
+		log.Error("更新失败!", err)
+		return
+	}
+	err, _ = UpdateFromStream(file)
+	fmt.Println()
+	if err != nil {
+		log.Error("更新失败!", err)
+		return
+	}
+	log.Info("更新完成！")
+}
